--- src/components/Feed/PostComposer/PostComposerModal/index.module.scss
@@ -1,217 +1,0 @@
-.modalContent {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-
-  background: #212121;
-  border-radius: 40px;
-
-  gap: 10px;
-  padding: 35px 30px;
-
-  @media (max-width: 768px) {
-    padding: 20px 15px;
-  }
-
-  .modalHeader {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-
-    width: 100%;
-
-    @media (max-width: 768px) {
-      width: 95%;
-    }
-
-    p {
-      font-size: 1.3rem;
-      font-weight: bold;
-    }
-
-    svg {
-      font-size: 1.3rem;
-      &:hover {
-        cursor: pointer;
-        color: #6c6c6c;
-      }
-    }
-  }
-
-  div.input {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-end;
-
-    textarea {
-      color: #000 !important;
-      background: #d9d9d9;
-      border: none;
-      border-radius: 18px;
-
-      padding: 10px 20px;
-
-      min-width: 500px;
-      resize: none;
-
-      min-height: 50px;
-
-      overflow: hidden;
-
-      @media (max-width: 768px) {
-        min-width: 70px;
-        min-height: 70px;
-
-        width: 99%;
-      }
-
-      &,
-      &::placeholder {
-        color: #7a7a7a;
-        font-weight: bold;
-        font-size: 1.3rem;
-
-        @media (max-width: 768px) {
-          font-size: 1.2rem;
-        }
-      }
-    }
-
-    span {
-      margin-top: 5px;
-      font-weight: bold;
-      transition: 1s;
-    }
-  }
-
-  .footer {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-
-    .buttons {
-      display: flex;
-      align-items: center;
-      gap: 5px;
-
-      button {
-        border: none;
-        border-radius: 30px;
-        background: #447ae3;
-        padding: 10px 40px;
-        font-size: 1.3rem;
-        font-weight: bold;
-        transition: 0.2s;
-
-        &:hover {
-          cursor: pointer;
-          background: #2e5bb8;
-        }
-
-        &[disabled] {
-          background: #6c6c6c;
-          &:hover {
-            cursor: not-allowed;
-          }
-        }
-      }
-
-      .fileInputLabel {
-        &:hover {
-          cursor: pointer;
-          color: #6c6c6c;
-        }
-
-        &[data-disabled="true"] {
-          color: #6c6c6c;
-          &:hover {
-            cursor: not-allowed;
-          }
-        }
-
-        svg {
-          font-size: 2rem;
-        }
-      }
-    }
-
-    p.publishingAs {
-      font-size: 1.2rem;
-      font-weight: bold;
-      margin-top: 10px;
-
-      user-select: none;
-
-      a {
-        color: #6c6c6c;
-        text-decoration: none;
-        &:hover {
-          cursor: pointer;
-          color: #447ae3;
-        }
-      }
-    }
-  }
-
-  .fileUploader {
-    .fileInputLabel {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-
-      width: 100%;
-      height: 100%;
-
-      border: 2px dashed #6c6c6c;
-      border-radius: 10px;
-
-      padding: 15px;
-
-      transition: 0.2s;
-
-      &:hover {
-        cursor: pointer;
-        border: 2px dashed #447ae3;
-      }
-
-      p {
-        font-size: 1.3rem;
-        font-weight: bold;
-        color: #929292;
-      }
-    }
-
-    .filesContainer {
-      display: flex;
-      gap: 10px;
-
-      @media (max-width: 768px) {
-        display: grid;
-        grid-template-columns: repeat(2, 1fr);
-      }
-
-      .filePreview {
-        svg {
-          display: block;
-          margin-bottom: 7px;
-          color: tomato;
-
-          &:hover {
-            cursor: pointer;
-            color: red;
-          }
-        }
-
-        img,
-        video {
-          object-fit: cover;
-          max-width: 200px;
-          max-height: 150px;
-          border-radius: 10px;
-        }
-      }
-    }
-  }
-}
