--- src/components/Modal/index.module.scss
@@ -1,12 +1,0 @@
-.modalContainer {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.76);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  z-index: 100;
-}
